sinew/sinew/requirements.txt
arrow==1.2.2
asgiref==3.4.1
autopep8==1.6.0
blessed==1.19.1
Django==4.0
django-cors-headers==3.10.1
django-picklefield==3.0.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
wcwidth==0.2.5
whitenoise==5.3.0
